Add protection against buffers filling up
Especially when dealing with larger streamed data, like in Root class for example, data may be produced much faster than it can be consumed by the client.
Usually it is not necessary (and worth) to yield the returned Promise by Response::stream() on small data, but one definitely should yield it on large data.

Also, Response methods only return $this on building operations now, not on sending ops (flush/send/end/stream).