Fix #124 - I hereby declare that I know what I'm doing
Due to vhosts now being explicitly keyed with the interface (and TLS being bound to a specific interface), it is not possible to get an unencrypted vhost on an encrypted connection and vice versa.

It is now permitted to set a specific expected port the client must specify in the Host header or wildcard port ("*") to avoid any port matching there.