Implemented listenAndServe, handleFunc and html functions
Added a couple of functions inspired in golang's http package.

`listenAndServe` makes really quick and straightforward to boot up a server passing a handler and a socket address.

`handleFunc` is a convenience function to transform a callable into a handler. It is less verbose that constructing the adapter.

`html` is just a wrapper for creating an html response.

Overall, these two functions make easier the quickstart process and make it a bit more elegant, redeable and a bit more "golangish".